Finished second task

+++ springboot/src/main/java/com/example/demo/entity/Country.java
@@ -7,6 +7,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.sql.Date;
+import java.util.List;
+import java.util.Set;
 
 @Setter
 @Getter
@@ -19,7 +21,7 @@ public class Country {
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
 //    @SequenceGenerator(name = "country_id", sequenceName = "id_sequence", initialValue = 1)
     @Column(name = "country_id")
-    private Long id;
+    private int id;
     private String name;
     private double area;
     @Column(name = "national_day")
@@ -30,4 +32,7 @@ public class Country {
     private String countryCode3;
 //    @Column(name = "region_id")
 //    private Date region;
+
+//    @OneToMany(cascade = CascadeType.ALL)
+//    private List<CountryStat> countryStats;
 }
